Initial Implementation 1
First part of getting the application ready for assessment additions.
Moved camera code to own camera class and added functionality:
Vertical Movement with keys and
Yaw and Pitch movement with mouse.
(Want to expand this to only move when mouse is clicked)
Updated Tools and Input classes to process more key and mouse inputs for extra camera control.