Object Inspector
Updated Mouse Picker selection to function with multiple selections using the CTRL key.
Implemented return of in game objects up the pipeline.
Implemented the first part of the Object Inspector:
Designed the dialogue with input boxes and sliders for tranform information.
Object ID and Transform information is displayed when selecting a single object.